requirements/test.txt
-r prod.txt
tox
coverage
pytest
pytest-cov
pytest-flask
docker-compose